[ALPS02291603] DISP:Fix coding defect
[Detail]
Need to fix coding defect in display driver.

[Solution]
Fix ddp_debug high impact coding defect.
Fix other low & medium impact coding defect.

[Feature] Display Quality Enhancement

MTK-Commit-Id: 972cfa6164b1351d9a28f215837cb35fe70cf407

CR-Id: ALPS02291603

+++ drivers/misc/mediatek/video/mt6735/dispsys/ddp_debug.h
@@ -9,8 +9,8 @@ extern unsigned int gResetRDMAEnable;
 extern unsigned int gOVLBackground;
 extern unsigned int gEnableIRQ;
 extern unsigned int gUltraEnable;
-extern unsigned int gRDMAUltraSetting;
-extern unsigned int gRDMAFIFOLen;
+extern unsigned long int gRDMAUltraSetting;
+extern unsigned long int gRDMAFIFOLen;
 extern unsigned int g_mobilelog;
 
 extern unsigned int disp_low_power_enlarge_blanking;
